Reorganize test infrastructure
- Slim global conftest.py (714 -> 45 lines): only core parametrization
  fixtures (voxel_size, image_shape, chunk_size, shared_tmpdir)
- Move all data fixtures, zarr writing, and CSV setup to
  tests/operations/conftest.py
- Centralize test helper functions in tests/test_utils.py, fixing the
  circular import (conftest importing from test_measure.py)
- Remove ~250 lines of duplicated helper code from test_measure.py
