+++ tests/conftest.py
@@ -1,14 +1,5 @@
 import pytest
 import numpy as np
-from funlib.geometry import Roi
-from skimage import measure
-from cellmap_analyze.util.zarr_util import create_multiscale_dataset
-import os
-from scipy import ndimage
-import pandas as pd
-from tests.operations.test_measure import simple_object_information_dict
-from funlib.persistence import prepare_ds
-from tests.contact_site_fixture_helper import compute_contact_sites_ground_truth
 
 
 @pytest.fixture(
@@ -39,26 +30,6 @@ def non_integer_voxel_sizes(request):
 
 
 @pytest.fixture(scope="session")
-def horizontal_cylinder_endpoints(voxel_size):
-    # Element-wise multiplication for anisotropic support
-    coords = np.array([[46, 3, 3], [46, 3, 40]], dtype=float)
-    # Scale by voxel_size per-axis and add half-voxel offset
-    return coords * voxel_size + 0.5 * voxel_size
-
-
-@pytest.fixture(scope="session")
-def vertical_cylinder_endpoints(voxel_size):
-    coords = np.array([[2, 2, 2], [2, 48, 2]], dtype=float)
-    return coords * voxel_size + 0.5 * voxel_size
-
-
-@pytest.fixture(scope="session")
-def diagonal_cylinder_endpoints(voxel_size):
-    coords = np.array([[45, 45, 5], [5, 5, 45]], dtype=float)
-    return coords * voxel_size + 0.5 * voxel_size
-
-
-@pytest.fixture(scope="session")
 def image_shape():
     return np.array((11, 11, 11))
 
@@ -69,646 +40,6 @@ def chunk_size():
 
 
 @pytest.fixture(scope="session")
-def image_with_holes(image_shape):
-    seg = np.zeros(image_shape, dtype=np.uint8)
-
-    # corner cube
-    seg[:3, :3, :3] = 1
-    seg[0, 0, 0] = 0  # shouldnt be filled since it is on border
-    seg[1, 1, 1] = 0  # should be filled
-
-    # hole spanning chunk
-    seg[3:7, 3:7, 3:7] = 2
-    seg[4:6, 4:6, 4:6] = 0
-
-    # two rectangles on top of eachother
-    seg[7:11, 7:11, 7:9] = 3
-    seg[7:11, 7:11, 9:11] = 4
-    seg[8:10, 8:10, 8:10] = 0  # hole between objects should be kept as a hole
-
-    return seg
-
-
-@pytest.fixture(scope="session")
-def random_image_to_delete(image_shape):
-    np.random.seed(42)
-    return np.random.randint(low=0, high=255, size=image_shape, dtype=np.uint8)
-
-
-@pytest.fixture(scope="session")
-def tic_tac_toe(image_shape):
-    seg = np.zeros(image_shape, dtype=np.uint16)
-    id = 2**8 + 1000
-    # Set the two slices along axis-0 (i.e., seg[3, :, :] and seg[7, :, :]) to 1
-    seg[3, :, :] = id
-    seg[7, :, :] = id
-
-    # Set the two slices along axis-1 (i.e., A[:, 3, :] and A[:, 7, :]) to 1
-    seg[:, 3, :] = id
-    seg[:, 7, :] = id
-
-    # Set the two slices along axis-2 (i.e., seg[:, :, 3] and seg[:, :, 7]) to 1
-    seg[:, :, 3] = id
-    seg[:, :, 7] = id
-
-    return seg
-
-
-@pytest.fixture(scope="session")
-def duplicate_ids(image_shape):
-    seg = np.zeros(image_shape, dtype=np.uint8)
-    seg[:7, :7, :7] = 1
-    seg[:7, 7:9, 7:9] = 2
-    seg[:7, 7:11, 9:11] = 1
-
-    return seg
-
-
-@pytest.fixture(scope="session")
-def no_duplicate_ids(image_shape):
-    seg = np.zeros(image_shape, dtype=np.uint8)
-    seg[:7, :7, :7] = 1
-    seg[:7, 7:9, 7:9] = 2
-    seg[:7, 7:11, 9:11] = 3
-
-    return seg
-
-
-@pytest.fixture(scope="session")
-def binarizable_image(image_shape):
-    seg = np.zeros(image_shape, dtype=np.uint8)
-    seg[0, 0, 0] = 1
-    seg[0, 1, 0] = 2
-    seg[0, 1:, 1] = 1
-    return seg
-
-
-@pytest.fixture(scope="session")
-def mask_one(image_shape):
-    mask = np.zeros(image_shape, dtype=np.uint8)
-    mask[:3, :3, :3] = 1
-    mask[4:6, 2:3, 4:6] = 2
-    return mask
-
-
-@pytest.fixture(scope="session")
-def mask_two(image_shape):
-    mask = np.zeros(image_shape, dtype=np.uint8)
-    mask[7:11, 7:11, 7:9] = 1
-    mask[7:11, 7:11, 9:11] = 2
-    return mask
-
-
-@pytest.fixture(scope="session")
-def label_mask(image_shape):
-    mask = np.zeros(image_shape, dtype=np.uint64)
-    mask[:5, :5, :5] = 1_000_000_000
-    mask[6:, 6:, :] = 2_000_000_0000
-    return mask
-
-
-@pytest.fixture(scope="session")
-def image_with_holes_filled(image_with_holes):
-    filled = np.zeros_like(image_with_holes)
-    for id in np.unique(image_with_holes[image_with_holes > 0]):
-        filled += (
-            ndimage.binary_fill_holes(image_with_holes == id).astype(np.uint8)
-        ) * id
-    return filled
-
-
-@pytest.fixture(scope="session")
-def tic_tac_toe_filled(tic_tac_toe):
-    seg = ndimage.binary_fill_holes(tic_tac_toe).astype(np.uint16)
-    seg[seg > 0] = 2**8 + 1000
-    return seg
-
-
-@pytest.fixture(scope="session")
-def blockwise_connected_components(image_shape):
-    seg = np.zeros(image_shape, dtype=np.uint64)
-    # single voxel
-    seg[1, 1, 1] = 1
-
-    # cube around border
-    id = 2
-    for x in range(3, 5):
-        for y in range(3, 5):
-            for z in range(3, 5):
-                seg[x, y, z] = id
-                id += 1
-
-    seg[7:11, 7:11, 7:11] = id + 1
-    return seg
-
-
-@pytest.fixture(scope="session")
-def intensity_image(image_shape):
-    seg = np.zeros(image_shape, dtype=np.uint8)
-    # single voxel
-    seg[1, 1, 1] = 127
-
-    # cube around border
-    for x in range(3, 5):
-        for y in range(3, 5):
-            for z in range(3, 5):
-                seg[x, y, z] = 127
-
-    seg[7:11, 7:11, 7:11] = 127
-    return seg
-
-
-@pytest.fixture(scope="session")
-def connected_components(intensity_image):
-    seg = measure.label(intensity_image > 0, connectivity=1)
-    return seg
-
-
-@pytest.fixture(scope="session")
-def segmentation_cylinders(
-    horizontal_cylinder_endpoints,
-    vertical_cylinder_endpoints,
-    diagonal_cylinder_endpoints,
-    voxel_size,
-):
-    def fill_in_cylinder(seg, endpoints_voxel, radius_nm, id, voxel_size):
-        # subtract 0.5 so that an annotation centered at a voxel matches to eg 0,0,0
-        end_2_voxel = endpoints_voxel[0] - 0.5
-        end_1_voxel = endpoints_voxel[1] - 0.5
-
-        # convert to physical space for distance computations
-        end_2 = end_2_voxel * voxel_size
-        end_1 = end_1_voxel * voxel_size
-
-        # https://stackoverflow.com/questions/56463412/distance-from-a-point-to-a-line-segment-in-3d-python
-        # normalized tangent vector in physical space
-        d = np.divide(end_2 - end_1, np.linalg.norm(end_2 - end_1))
-
-        # possible points in voxel space
-        radius_voxels = np.ceil(radius_nm / voxel_size).astype(int)
-        mins = np.floor(np.minimum(end_1_voxel, end_2_voxel)).astype(int) - (
-            radius_voxels + 1
-        )  # 1s for padding
-        maxs = np.ceil(np.maximum(end_1_voxel, end_2_voxel)).astype(int) + (
-            radius_voxels + 1
-        )
-
-        z, y, x = [list(range(mins[i], maxs[i] + 1, 1)) for i in range(3)]
-        p_voxel = np.array(np.meshgrid(z, y, x)).T.reshape((-1, 3))
-
-        # convert candidate points to physical space
-        p = p_voxel * voxel_size
-
-        # signed parallel distance components in physical space
-        s = np.dot(end_1 - p, d)
-        t = np.dot(p - end_2, d)
-
-        # clamped parallel distance
-        h = np.maximum.reduce([s, t, np.zeros_like(s)])
-
-        # perpendicular distance component in physical space
-        c = np.linalg.norm(np.cross(p - end_1, d), axis=1)
-
-        is_in_cylinder = (h == 0) & (c <= radius_nm)
-        voxels_in_cylinder = p_voxel[is_in_cylinder]
-        seg[
-            voxels_in_cylinder[:, 0], voxels_in_cylinder[:, 1], voxels_in_cylinder[:, 2]
-        ] = id
-
-    # need to use larger to get better chance of fitting
-    seg = np.zeros((50, 50, 50), dtype=np.uint8)
-    min_voxel = np.min(voxel_size)
-    # horizontal cylinder
-    fill_in_cylinder(seg, horizontal_cylinder_endpoints / voxel_size, 1 * min_voxel, 1, voxel_size)
-    seg[0, 0, 0] = 1  # add spurious voxel
-    # vertical cylinder
-    fill_in_cylinder(seg, vertical_cylinder_endpoints / voxel_size, 1.5 * min_voxel, 2, voxel_size)
-    # diagonal
-    fill_in_cylinder(seg, diagonal_cylinder_endpoints / voxel_size, 2 * min_voxel, 3, voxel_size)
-
-    return seg
-
-
-@pytest.fixture(scope="session")
-def segmentation_spheres(
-    shape=(50, 50, 50),
-    spheres=[((12, 12, 12), 5), ((25, 25, 25), 20), ((40, 40, 40), 8)],
-):
-    seg = np.zeros(shape, dtype=np.uint8)
-    z, y, x = np.indices(shape)
-
-    for label, ((zc, yc, xc), r) in enumerate(spheres, start=1):
-        mask = (z - zc) ** 2 + (y - yc) ** 2 + (x - xc) ** 2 < r**2
-        seg[mask] = label
-    seg[0:6, 15:36, 22:29] = 2
-    return seg
-
-
-@pytest.fixture(scope="session")
-def affinities_offsets():
-    return [
-        [1, 0, 0],
-        [0, 1, 0],
-        [0, 0, 1],
-        [3, 0, 0],
-        [0, 3, 0],
-        [0, 0, 3],
-        [9, 0, 0],
-        [0, 9, 0],
-        [0, 0, 9],
-    ]
-
-
-@pytest.fixture(scope="session")
-def affinities_cylinders(segmentation_cylinders, affinities_offsets):
-    """
-    Create a new array with shape (num_offsets, *seg.shape) where each channel
-    is 1 if the pixel at the offset from the current pixel belongs to the same object,
-    and 0 otherwise.
-
-    Parameters:
-      seg: numpy array of shape (D, H, W) containing segmentation labels.
-      offsets: list of tuples, e.g. [(1,0,0), (0,1,0), ...] representing the offsets.
-
-    Returns:
-      out: numpy array of shape (len(offsets), D, H, W)
-    """
-    shape = segmentation_cylinders.shape
-    num_offsets = len(affinities_offsets)
-    out = np.zeros((num_offsets,) + shape, dtype=np.uint8)
-
-    for idx, (dx, dy, dz) in enumerate(affinities_offsets):
-        # Create an array for the shifted segmentation; fill with -1 to mark invalid areas.
-        shifted = np.zeros(shape, dtype=segmentation_cylinders.dtype)
-
-        # Compute slices for the valid region in the source and destination.
-        src_slice = [
-            slice(max(dx, 0), shape[0] + min(dx, 0)),
-            slice(max(dy, 0), shape[1] + min(dy, 0)),
-            slice(max(dz, 0), shape[2] + min(dz, 0)),
-        ]
-
-        dst_slice = [
-            slice(max(-dx, 0), shape[0] + min(-dx, 0)),
-            slice(max(-dy, 0), shape[1] + min(-dy, 0)),
-            slice(max(-dz, 0), shape[2] + min(-dz, 0)),
-        ]
-
-        # Shift the segmentation array.
-        shifted[tuple(dst_slice)] = segmentation_cylinders[tuple(src_slice)]
-
-        # For each pixel, assign 1 if the label matches that of the shifted pixel.
-        out[idx] = (
-            (segmentation_cylinders == shifted).astype(np.uint8)
-            * (segmentation_cylinders > 0)
-            * 255
-        )
-    return out
-
-
-@pytest.fixture(autouse=True, scope="session")
-def segmentation_cells(image_shape):
-    seg = np.zeros(image_shape, dtype=np.uint8)
-    # assign different values to different corners
-    seg[:2, :2, :2] = 1
-    seg[-2:, -2:, -2:] = 2
-    seg[5:8, 5:8, 5:8] = 3
-    return seg
-
-
-@pytest.fixture(autouse=True, scope="session")
-def tmp_coms_csv(shared_tmpdir, voxel_size):
-    output_path = shared_tmpdir + "/csvs/"
-    os.makedirs(name=output_path, exist_ok=True)
-    df = pd.DataFrame(
-        {
-            "Object ID": [1, 2, 3],
-            "COM X (nm)": (np.array([1.1, 3.1, 11.1]) + 0.5) * voxel_size[2],  # X
-            "COM Y (nm)": (np.array([1.1, 3.1, 11.1]) + 0.5) * voxel_size[1],  # Y
-            "COM Z (nm)": (np.array([1.1, 3.1, 8.1]) + 0.5) * voxel_size[0],  # Z
-        }
-    )
-    df.to_csv(output_path + "/assignment_coms.csv", index=False)
-    return str(output_path + "/assignment_coms.csv")
-
-
-@pytest.fixture(autouse=True, scope="session")
-def tmp_cylinders_information_csv(segmentation_cylinders, shared_tmpdir, voxel_size):
-    output_path = shared_tmpdir + "/csvs/"
-    os.makedirs(name=output_path, exist_ok=True)
-    cylinder_information_dict = simple_object_information_dict(
-        segmentation_cylinders, voxel_size
-    )
-    # create dataframe
-    columns = [
-        "Object ID",
-        "Volume (nm^3)",
-        "Surface Area (nm^2)",
-        "Radius of Gyration (nm)",
-    ]
-    for category in ["COM", "MIN", "MAX"]:
-        for d in ["X", "Y", "Z"]:
-            columns.append(f"{category} {d} (nm)")
-    df = pd.DataFrame(
-        index=np.arange(len(cylinder_information_dict)),
-        columns=columns,
-    )
-    for i, (id, oi) in enumerate(cylinder_information_dict.items()):
-        row = [
-            id,
-            float(oi.volume),  # Ensure volume is float to match actual output
-            float(
-                oi.surface_area
-            ),  # Ensure surface_area is float to match actual output
-            oi.radius_of_gyration,
-            *oi.com[::-1],
-            *oi.bounding_box[:3][::-1],
-            *oi.bounding_box[3:][::-1],
-        ]
-        df.loc[i] = row
-
-    df.to_csv(output_path + "/segmentation_cylinders.csv", index=False)
-    return str(output_path + "/segmentation_cylinders.csv")
-
-
-@pytest.fixture(scope="session")
-def segmentation_1(image_shape):
-    seg = np.zeros(image_shape, dtype=np.uint8)
-    seg[:2, 2:10, 2:10] = 1
-    return seg
-
-
-@pytest.fixture(scope="session")
-def segmentation_1_downsampled(segmentation_1):
-    # take every other pixel
-    return segmentation_1[::2, ::2, ::2]
-
-
-@pytest.fixture(scope="session")
-def segmentation_2(image_shape):
-    seg = np.zeros(image_shape, dtype=np.uint8)
-    seg[3:5, 2:5, 4:6] = 1
-    seg[3:9, 7:10, 4:8] = 2
-    return seg
-
-
-@pytest.fixture(scope="session")
-def segmentation_random(image_shape):
-    np.random.seed(42)
-    seg = np.random.randint(low=1, high=10, size=image_shape, dtype=np.uint8)
-    return seg
-
-
-@pytest.fixture(scope="session")
-def raw_intensity_for_seg2(image_shape):
-    np.random.seed(123)
-    return np.random.rand(*image_shape).astype(np.float32) * 255
-
-
-@pytest.fixture(scope="session")
-def contact_sites_distance_8nm(segmentation_1, segmentation_2, voxel_size):
-    """Ground truth contact sites for contact_distance_nm=8."""
-    return compute_contact_sites_ground_truth(
-        segmentation_1, segmentation_2, contact_distance_nm=8, voxel_size=voxel_size
-    )
-
-
-@pytest.fixture(scope="session")
-def contact_sites_distance_16nm(segmentation_1, segmentation_2, voxel_size):
-    """Ground truth contact sites for contact_distance_nm=16."""
-    return compute_contact_sites_ground_truth(
-        segmentation_1, segmentation_2, contact_distance_nm=16, voxel_size=voxel_size
-    )
-
-
-@pytest.fixture(scope="session")
-def contact_sites_distance_24nm(segmentation_1, segmentation_2, voxel_size):
-    """Ground truth contact sites for contact_distance_nm=24."""
-    return compute_contact_sites_ground_truth(
-        segmentation_1, segmentation_2, contact_distance_nm=24, voxel_size=voxel_size
-    )
-
-
-import pytest
-
-
-@pytest.fixture(scope="session")
 def shared_tmpdir(tmpdir_factory):
     """Create a shared temporary directory for all test functions."""
     return tmpdir_factory.mktemp("tmp")
-
-
-@pytest.fixture(scope="session")
-def tmp_zarr(shared_tmpdir, voxel_size):
-    # Create separate zarr files for each voxel_size parametrization
-    voxel_size_str = f"{voxel_size[0]}_{voxel_size[1]}_{voxel_size[2]}"
-    output_path = shared_tmpdir + f"/tmp_{voxel_size_str}.zarr"
-    os.makedirs(name=output_path, exist_ok=True)
-    return str(output_path)
-
-
-@pytest.fixture(autouse=True, scope="session")
-def tmp_object_information_csv(shared_tmpdir):
-    # do it this way otherwise it appends 0 to the end of the zarr
-    output_path = shared_tmpdir + "/csvs/"
-    os.makedirs(name=output_path, exist_ok=True)
-    df = pd.DataFrame(
-        {
-            "Object ID": [1, 4],
-            "Volume (nm^3)": [100, 400],
-            "Surface Area (nm^2)": [100, 400],
-        }
-    )
-    df.to_csv(output_path + "/objects.csv", index=False)
-    return str(output_path + "/objects.csv")
-
-
-@pytest.fixture(scope="session")
-def segmentation_for_skeleton():
-    """Create a segmentation with multiple distinct IDs of various shapes for skeletonization tests."""
-    # Use larger shape for more complex structures
-    seg = np.zeros((50, 50, 50), dtype=np.uint8)
-
-    # ID 1: Small cube in corner
-    seg[1:4, 1:4, 1:4] = 1
-
-    # ID 2: Elongated structure (good for skeletonization)
-    seg[5:15, 2:5, 2:5] = 2
-
-    # ID 3: Another elongated structure (perpendicular to ID 2)
-    seg[2:5, 5:15, 2:5] = 3
-
-    # ID 4: Sphere (should skeletonize to a point or small cluster)
-    center = (25, 25, 25)
-    radius = 4
-    z, y, x = np.ogrid[0:50, 0:50, 0:50]
-    sphere_mask = (z - center[0]) ** 2 + (y - center[1]) ** 2 + (
-        x - center[2]
-    ) ** 2 <= radius**2
-    seg[sphere_mask] = 4
-
-    # ID 5: Cross/plus shape (3D cross with branches in Z, X, Y directions)
-    # Thick arms (4x4 cross-section) so skeletonize produces real branches.
-    # Z arm (top/bottom of junction): z=30..49, 20 voxels long
-    seg[30:50, 23:27, 23:27] = 5
-    # X arm (longest): z=38..41, x=10..39, 30 voxels long
-    seg[38:42, 23:27, 10:40] = 5
-    # Y arm: z=38..41, y=14..35, 22 voxels long
-    seg[38:42, 14:36, 23:27] = 5
-
-    # ID 6: L-shaped structure
-    seg[5:10, 30:35, 30:32] = 6  # Vertical part
-    seg[8:10, 30:32, 30:40] = 6  # Horizontal part
-
-    # ID 7: Figure-8 shape (two loops in a single z-plane)
-    z, y, x = np.ogrid[0:50, 0:50, 0:50]
-
-    # Fix the plane for the figure 8
-    z_plane = 15
-    plane_mask = z == z_plane
-
-    # Loop parameters
-    bottom_center = (45, 15)  # (y, x)
-    top_center = (45, 23)
-    radius = 3
-
-    # Bottom loop
-    bottom_ring = (
-        plane_mask
-        & (
-            (y - bottom_center[0]) ** 2 + (x - bottom_center[1]) ** 2
-            >= (radius - 0.5) ** 2
-        )
-        & (
-            (y - bottom_center[0]) ** 2 + (x - bottom_center[1]) ** 2
-            <= (radius + 0.5) ** 2
-        )
-    )
-    seg[bottom_ring] = 7
-
-    # Top loop
-    top_ring = (
-        plane_mask
-        & ((y - top_center[0]) ** 2 + (x - top_center[1]) ** 2 >= (radius - 0.5) ** 2)
-        & ((y - top_center[0]) ** 2 + (x - top_center[1]) ** 2 <= (radius + 0.5) ** 2)
-    )
-    seg[top_ring] = 7
-
-    # The connection (branch point) — also in a single plane
-    seg[z_plane, 44:47, 18:20] = 7
-
-    return seg
-
-
-@pytest.fixture(scope="session")
-def test_image_dict(
-    blockwise_connected_components,
-    connected_components,
-    intensity_image,
-    image_with_holes,
-    image_with_holes_filled,
-    tic_tac_toe,
-    tic_tac_toe_filled,
-    no_duplicate_ids,
-    duplicate_ids,
-    binarizable_image,
-    mask_one,
-    mask_two,
-    label_mask,
-    random_image_to_delete,
-    segmentation_1,
-    segmentation_2,
-    segmentation_random,
-    segmentation_cylinders,
-    segmentation_spheres,
-    segmentation_cells,
-    affinities_cylinders,
-    segmentation_1_downsampled,
-    contact_sites_distance_8nm,
-    contact_sites_distance_16nm,
-    contact_sites_distance_24nm,
-    segmentation_for_skeleton,
-    raw_intensity_for_seg2,
-):
-    dict = {
-        "blockwise_connected_components": blockwise_connected_components,
-        "connected_components": connected_components,
-        "intensity_image": intensity_image,
-        "image_with_holes": image_with_holes,
-        "image_with_holes_filled": image_with_holes_filled,
-        "tic_tac_toe": tic_tac_toe,
-        "tic_tac_toe_filled": tic_tac_toe_filled,
-        "binarizable_image": binarizable_image,
-        "duplicate_ids": duplicate_ids,
-        "no_duplicate_ids": no_duplicate_ids,
-        "mask_one": mask_one,
-        "mask_two": mask_two,
-        "label_mask": label_mask,
-        "random_image_to_delete": random_image_to_delete,
-        "segmentation_1": segmentation_1,
-        "segmentation_cylinders": segmentation_cylinders,
-        "segmentation_spheres": segmentation_spheres,
-        "affinities_cylinders": affinities_cylinders,
-        "segmentation_1_downsampled": segmentation_1_downsampled,
-        "segmentation_2": segmentation_2,
-        "segmentation_random": segmentation_random,
-        "segmentation_cells": segmentation_cells,
-        "segmentation_for_skeleton": segmentation_for_skeleton,
-        "raw_intensity_for_seg2": raw_intensity_for_seg2,
-        "contact_sites_distance_8nm": contact_sites_distance_8nm,
-        "contact_sites_distance_16nm": contact_sites_distance_16nm,
-        "contact_sites_distance_24nm": contact_sites_distance_24nm,
-    }
-
-    return dict
-
-
-@pytest.fixture(autouse=True, scope="session")
-def write_zarrs(tmp_zarr, test_image_dict, voxel_size, chunk_size):
-    for data_name, data in test_image_dict.items():
-        # Handle both tuple and scalar voxel_size
-        current_voxel_size = (
-            voxel_size if "downsampled" not in data_name else 2 * voxel_size
-        )
-        data_path = f"{tmp_zarr}/{data_name}"
-
-        # Use the voxel_size tuple directly (already in (Z, Y, X) format)
-        # Convert to Python ints for JSON serialization
-        ds_voxel_size = tuple(int(v) for v in current_voxel_size)
-        data_shape = data.shape
-        if len(data_shape) == 4:
-            data_shape = data_shape[1:]
-
-        total_roi = Roi((0, 0, 0), np.array(data_shape) * current_voxel_size)
-        write_size = (
-            chunk_size * current_voxel_size
-            if (
-                data_name != "segmentation_cylinders"
-                and data_name != "affinities_cylinders"
-                and data_name != "segmentation_spheres"
-                and data_name != "segmentation_for_skeleton"
-            )
-            else np.array((20, 20, 20)) * current_voxel_size
-        )
-
-        if "affinities" in data_name:
-            ds = prepare_ds(
-                tmp_zarr,
-                data_name,
-                total_roi=total_roi,
-                write_size=write_size,
-                voxel_size=ds_voxel_size,
-                dtype=np.uint8,
-                num_channels=np.shape(data)[0],
-            )
-        else:
-            # use larger chunk size for segmentation_cylinders since we had to make it bigger
-            ds = create_multiscale_dataset(
-                data_path,
-                dtype=data.dtype,
-                voxel_size=ds_voxel_size,
-                total_roi=total_roi,
-                write_size=write_size,
-            )
-
-        ds.data[:] = data
